refactor(team): extract wiki worker lifecycle from broker.go
Phase 3 third slice. Pulls ensureWikiWorker (96 LOC) into a kin file.
This is the entire markdown-backend cold start: repo init, fsck +
backup-restore double-fault recovery, index construction, worker +
extractor + DLQ wiring, skill-status disk reconciliation, async boot
reconcile, lint cron, skill-compile cron, Stage B synthesizer. One
file end-to-end now, instead of buried mid-Start().

  - internal/team/broker.go: -100 LOC (2800 -> 2700)
  - internal/team/broker_wiki_lifecycle.go: +118 LOC (new)

Cumulative broker.go reduction across Phase 3:
  3404 -> 2859 (web proxy + sse)         -545
       -> 2800 (auth + publish)          -59
       -> 2700 (wiki lifecycle)          -100
  Net: -704 LOC; broker.go shrunk by 21%.

No behavior change. Tests pass.
